v0-0-1: BUG - CACU is hanging for the first packet
If the display filter is not met, CACU never enters the for loop because there's nothing to iterate over.  There's no clean way to exit out of CACU if it never detects a single packet matching the display filter.

I'm about to try something kinda drastic in order to avoid hanging while waiting on the first packet.