v0-0-1: Fixed "hanging" BUG in CACU
Cautious Cudgel was "hanging" in situations in which there was no filtered traffic to be had.  To avoid that situation, I wrote a "peek" helper function to peek into the generator to determine if there were any packets to be read yet.  I put this in a while loop, also updating the user visually and giving the user the opportunity to 'abort'.

To that end:
	Added peek_into_generator()
	Extricated 'main' into an actual def main(): function
	Cleaned up some of the 'outlining' in the comments
	Modified the 'outline' of the logic flow
	Re(un)numbered some comments
	Implemented the new "peek" function within main()