v0-0-1: BUG HUNT non-gen into p_i_g()
Turns out I wasn't passing an actual generator into peek_into_generator().  Modified the  'live capture' use case call to p_i_g() to actually pass in the generator.  The returned generator is then assigned to packetGen when p_i_g() detects one.  Also, the for 'for packet in...' loop had to be changed since the actual generator (and not the method) is being stored in packetGen.